Initial Plugin restructure
