+++ plugins/model/Model_GAN/Trainer.py
