+++ plugins/model/Model_GAN/instance_normalization.py
