+++ plugins/model/Model_OriginalHighRes/instance_normalization.py
