+++ scripts/convert.py
@@ -12,7 +12,7 @@ from scripts.fsmedia import Alignments, Images, Faces, Utils
 from scripts.extract import Extract
 from lib.utils import BackgroundGenerator, get_folder, get_image_paths
 
-from plugins.PluginLoader import PluginLoader
+from plugins.plugin_loader import PluginLoader
 
 
 class Convert(object):
